Organize, lint, format
