+++ packages/web/src/components/pdf/index.js
@@ -3,7 +3,8 @@
  */
 
 // PDF list components (for displaying PDF metadata in lists)
-export { PdfListItem, PdfTagBadge } from './pdfjs/index.js';
+export { default as PdfListItem } from './PdfListItem.jsx';
+export { default as PdfTagBadge } from './PdfTagBadge.jsx';
 
 // EmbedPDF viewer component
 export { default as EmbedPdfViewer } from './embedpdf/EmbedPdfViewer.jsx';
