--- packages/web/src/components/pdf/pdfjs/index.js
@@ -1,7 +1,0 @@
-/**
- * PDF Components - Barrel export
- * Only PdfListItem and PdfTagBadge remain (used for displaying PDF metadata in lists)
- */
-
-export { default as PdfListItem } from './PdfListItem.jsx';
-export { default as PdfTagBadge } from './PdfTagBadge.jsx';
