+++ packages/web/src/components/project/all-studies-tab/study-card/StudyPdfSection.jsx
@@ -8,7 +8,7 @@
 import { createSignal, createMemo, Show, For } from 'solid-js';
 import { FaBrandsGoogleDrive, FaSolidPlus } from 'solid-icons/fa';
 import { showToast } from '@corates/ui';
-import PdfListItem from '@pdf/pdfjs/PdfListItem.jsx';
+import { PdfListItem } from '@pdf';
 import EditPdfMetadataModal from '@/components/project/all-studies-tab/EditPdfMetadataModal.jsx';
 import projectActionsStore from '@/stores/projectActionsStore';
