+++ packages/web/src/components/project/todo-tab/TodoStudyRow.jsx
@@ -9,7 +9,7 @@ import { For, Show, createMemo, createSignal } from 'solid-js';
 import { BiRegularChevronRight } from 'solid-icons/bi';
 import { Collapsible } from '@corates/ui';
 import { getChecklistMetadata } from '@/checklist-registry';
-import PdfListItem from '@pdf/pdfjs/PdfListItem.jsx';
+import { PdfListItem } from '@pdf';
 import ChecklistForm from './ChecklistForm.jsx';
 import { getStatusLabel, getStatusStyle } from '@/constants/checklist-status.js';
